requirements.txt

pdfminer.six
pandas